# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AfriPebbles API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /orders:
    post:
      operationId: createOrder
      tags: [orders]
      summary: Create a guest order. Prices/availability are re-validated and computed server-side; client-submitted amounts are never trusted.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderInput"
      responses:
        "201":
          description: Order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderWithItems"
        "400":
          description: Invalid cart or missing consent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /admin/orders:
    get:
      operationId: adminListOrders
      tags: [admin]
      summary: List orders with filters, search, and pagination
      parameters:
        - name: paymentStatus
          in: query
          schema:
            $ref: "#/components/schemas/PaymentStatus"
        - name: orderStatus
          in: query
          schema:
            $ref: "#/components/schemas/OrderStatus"
        - name: country
          in: query
          schema:
            type: string
        - name: dateFrom
          in: query
          schema:
            type: string
            format: date-time
        - name: dateTo
          in: query
          schema:
            type: string
            format: date-time
        - name: search
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: number
        - name: offset
          in: query
          schema:
            type: number
      responses:
        "200":
          description: Paginated order list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminOrdersPage"
  /admin/orders/{id}:
    get:
      operationId: adminGetOrder
      tags: [admin]
      summary: Get an order, its items, and its status-change history
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Order detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminOrderDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: adminUpdateOrder
      tags: [admin]
      summary: Update payment/order status, tracking number, or internal notes. Status changes are recorded to the audit trail.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminUpdateOrderInput"
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderWithItems"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ---------------------------------------------------------------------------
  # Admin: coupons
  # ---------------------------------------------------------------------------
  /admin/coupons:
    get:
      operationId: adminListCoupons
      tags: [admin]
      summary: List all coupons
      responses:
        "200":
          description: All coupons
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Coupon"
    post:
      operationId: adminCreateCoupon
      tags: [admin]
      summary: Create a coupon
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CouponInput"
      responses:
        "201":
          description: Coupon created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Coupon"
        "409":
          description: Code already in use
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/coupons/{id}:
    get:
      operationId: adminGetCoupon
      tags: [admin]
      summary: Get a coupon by id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Coupon
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Coupon"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: adminUpdateCoupon
      tags: [admin]
      summary: Update a coupon
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CouponInput"
      responses:
        "200":
          description: Coupon updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Coupon"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: adminDeleteCoupon
      tags: [admin]
      summary: Delete a coupon
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    PaymentMethod:
      type: string
      enum: [paypal, mobile_money, bank_transfer]
    PaymentStatus:
      type: string
      enum: [pending, paid, failed, refunded, cancelled]
    OrderStatus:
      type: string
      enum: [pending_payment, paid, processing, shipped, delivered, cancelled]
    CouponDiscountType:
      type: string
      enum: [percentage, fixed]
    OrderItemVariant:
      type: object
      properties:
        label:
          type: string
        option:
          type: string
      required:
        - label
        - option
    CartItemInput:
      type: object
      properties:
        productId:
          type: number
        quantity:
          type: number
        variant:
          oneOf:
            - $ref: "#/components/schemas/OrderItemVariant"
            - type: "null"
      required:
        - productId
        - quantity
    OrderItem:
      type: object
      properties:
        id:
          type: number
        orderId:
          type: number
        productId:
          type: ["number", "null"]
        productName:
          type: string
        productType:
          type: string
          enum: [digital, physical]
        variant:
          oneOf:
            - $ref: "#/components/schemas/OrderItemVariant"
            - type: "null"
        quantity:
          type: number
        unitPrice:
          type: number
        lineTotal:
          type: number
        shippingAmount:
          type: number
        isDigital:
          type: boolean
        isPreorder:
          type: boolean
        preorderFulfilmentText:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - orderId
        - productId
        - productName
        - productType
        - variant
        - quantity
        - unitPrice
        - lineTotal
        - shippingAmount
        - isDigital
        - isPreorder
        - preorderFulfilmentText
        - createdAt
    Order:
      type: object
      properties:
        id:
          type: number
        orderReference:
          type: string
        customerName:
          type: string
        customerEmail:
          type: string
        customerPhone:
          type: string
        country:
          type: string
        deliveryAddress:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        currency:
          type: string
        subtotal:
          type: number
        shippingTotal:
          type: number
        discountTotal:
          type: number
        grandTotal:
          type: number
        paymentMethod:
          $ref: "#/components/schemas/PaymentMethod"
        paymentStatus:
          $ref: "#/components/schemas/PaymentStatus"
        orderStatus:
          $ref: "#/components/schemas/OrderStatus"
        paymentNote:
          type: ["string", "null"]
        paidAt:
          type: ["string", "null"]
          format: date-time
        trackingNumber:
          type: ["string", "null"]
        internalNotes:
          type: ["string", "null"]
        couponCode:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - orderReference
        - customerName
        - customerEmail
        - customerPhone
        - country
        - deliveryAddress
        - notes
        - currency
        - subtotal
        - shippingTotal
        - discountTotal
        - grandTotal
        - paymentMethod
        - paymentStatus
        - orderStatus
        - paymentNote
        - paidAt
        - trackingNumber
        - internalNotes
        - couponCode
        - createdAt
        - updatedAt
    OrderWithItems:
      allOf:
        - $ref: "#/components/schemas/Order"
        - type: object
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/OrderItem"
          required:
            - items
    OrderStatusHistoryEntry:
      type: object
      properties:
        id:
          type: number
        orderId:
          type: number
        changedBy:
          type: string
        field:
          type: string
          enum: [payment_status, order_status]
        oldValue:
          type: string
        newValue:
          type: string
        note:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - orderId
        - changedBy
        - field
        - oldValue
        - newValue
        - note
        - createdAt
    AdminOrderDetail:
      type: object
      properties:
        order:
          $ref: "#/components/schemas/OrderWithItems"
        history:
          type: array
          items:
            $ref: "#/components/schemas/OrderStatusHistoryEntry"
      required:
        - order
        - history
    AdminOrdersPage:
      type: object
      properties:
        orders:
          type: array
          items:
            $ref: "#/components/schemas/Order"
        total:
          type: number
      required:
        - orders
        - total
    AdminUpdateOrderInput:
      type: object
      properties:
        paymentStatus:
          $ref: "#/components/schemas/PaymentStatus"
        orderStatus:
          $ref: "#/components/schemas/OrderStatus"
        trackingNumber:
          type: ["string", "null"]
        internalNotes:
          type: ["string", "null"]
        paymentNote:
          type: ["string", "null"]
        statusChangeNote:
          type: string
...
    CreateOrderInput:
      type: object
      properties:
        customerName:
          type: string
        customerEmail:
          type: string
        customerPhone:
          type: string
        country:
          type: string
        deliveryAddress:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        paymentMethod:
          $ref: "#/components/schemas/PaymentMethod"
        couponCode:
          type: ["string", "null"]
        items:
          type: array
          items:
            $ref: "#/components/schemas/CartItemInput"
        consent:
          type: boolean
      required:
        - customerName
        - customerEmail
        - customerPhone
        - country
        - paymentMethod
        - items
        - consent
...
    Coupon:
      type: object
      properties:
        id:
          type: number
        code:
          type: string
        discountType:
          $ref: "#/components/schemas/CouponDiscountType"
        discountValue:
          type: number
        currency:
          type: ["string", "null"]
        activeFrom:
          type: ["string", "null"]
          format: date-time
        activeUntil:
          type: ["string", "null"]
          format: date-time
        usageLimit:
          type: ["number", "null"]
        usageCount:
          type: number
        minimumOrderAmount:
          type: ["number", "null"]
        isActive:
          type: boolean
        restrictedProductIds:
          type: ["array", "null"]
          items:
            type: number
        restrictedCategory:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - code
        - discountType
        - discountValue
        - currency
        - activeFrom
        - activeUntil
        - usageLimit
        - usageCount
        - minimumOrderAmount
        - isActive
        - restrictedProductIds
        - restrictedCategory
        - createdAt
        - updatedAt
    CouponInput:
      type: object
      properties:
        code:
          type: string
        discountType:
          $ref: "#/components/schemas/CouponDiscountType"
        discountValue:
          type: number
        currency:
          type: ["string", "null"]
        activeFrom:
          type: ["string", "null"]
          format: date-time
        activeUntil:
          type: ["string", "null"]
          format: date-time
        usageLimit:
          type: ["number", "null"]
        minimumOrderAmount:
          type: ["number", "null"]
        isActive:
          type: boolean
        restrictedProductIds:
          type: ["array", "null"]
          items:
            type: number
        restrictedCategory:
          type: ["string", "null"]
      required:
        - code
        - discountType
        - discountValue