# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * AfriPebbles API specification
 * OpenAPI spec version: 0.1.0
 */
import type { OrderStatus } from './orderStatus';
import type { PaymentMethod } from './paymentMethod';
import type { PaymentStatus } from './paymentStatus';

export interface Order {
  id: number;
  orderReference: string;
  customerName: string;
  customerEmail: string;
  customerPhone: string;
  country: string;
  /** @nullable */
  deliveryAddress: string | null;
  /** @nullable */
  notes: string | null;
  currency: string;
  subtotal: number;
  shippingTotal: number;
  discountTotal: number;
  grandTotal: number;
  paymentMethod: PaymentMethod;
  paymentStatus: PaymentStatus;
  orderStatus: OrderStatus;
  /** @nullable */
  paymentNote: string | null;
  /** @nullable */
  paidAt: Date | null;
  /** @nullable */
  trackingNumber: string | null;
  /** @nullable */
  internalNotes: string | null;
  /** @nullable */
  couponCode: string | null;
  createdAt: Date;
  updatedAt: Date;
}