# lib/api-zod/src/generated/types/orderItemVariant.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * AfriPebbles API specification
 * OpenAPI spec version: 0.1.0
 */

export interface OrderItemVariant {
  label: string;
  option: string;
}