# lib/api-zod/src/generated/types/paymentMethod.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * AfriPebbles API specification
 * OpenAPI spec version: 0.1.0
 */

export type PaymentMethod = typeof PaymentMethod[keyof typeof PaymentMethod];


export const PaymentMethod = {
  paypal: 'paypal',
  mobile_money: 'mobile_money',
  bank_transfer: 'bank_transfer',
} as const;